Update mp3_convertor.sh
aff validate size of records

+++ aster/mp3_convertor.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/sh
 echo "starting......"
 #####################################################
 # docker-compose -f docker-compose.yml stop callrecordssheduler \
@@ -22,39 +22,49 @@ echo "######### Will Use system Variables for AWS S3"
 echo "######### AWSS3 bucket== $AWSS3"
 echo "#############################################"
 awss3bucket="$(echo "$AWSS3")"
-deltatime="5" # in minutes (-cmin)
+deltatime="3" # in minutes (-mmin)
 bitrate_to_mp3="32"
+requiredsizewavfile=100
 records_path="/tmp/callsrecords"
 wav_records="$records_path/wav"
-mp3_records="$records_path/mp3"
+datepath="$(date +"%Y")/$(date +"%m")/$(date +"%d")"
+mp3_records="$records_path/mp3/$datepath"
 
 echo "=================  awss3bucket   $awss3bucket "
 
 mkdir -p $wav_records
-mkdir -p $mp3_records
 
 ls -la $wav_records
 #ls -la $mp3_records
 
-list_wav_files=$(find $wav_records/* -cmin +$deltatime | sed 's/.*\///' | grep -v mp3)
-#list_mp3_files=$(find $mp3_records/* -cmin +$deltatime | sed 's/.*\///')
+list_wav_files=$(find $wav_records/* -mmin +$deltatime | sed 's/.*\///' | grep -v mp3)
+#list_mp3_files=$(find $mp3_records/* -mmin +$deltatime | sed 's/.*\///')
 
 echo $list_wav_files
 #echo $list_mp3_files
 
 echo "Start converting and moving to AWS S3 storage......................."
-for wav_file in  ${list_wav_files[*]}  
+for wav_file in $list_wav_files;
 do
   echo "----------------------"
   printf "   %s\n" $wav_records/$wav_file
   mp3_file="$(echo $wav_file | sed 's/.wav//g').mp3"
-  echo "name of MP3 file == $mp3_file -------------------------------------------------"
-  lame -b $bitrate_to_mp3 $wav_records/$wav_file $wav_records/$mp3_file
+  if [[ $(stat -c%s $wav_records/$wav_file) -lt $requiredsizewavfile ]]
+  then
+    echo "-------------- file $wav_records/$wav_file LESSSSSSSSSSSS then $requiredsizewavfile "
+  else
+    mkdir -p $mp3_records
+    echo "-------------- full way to mp3 file  -------- $mp3_records/$mp3_file-- ---------------------"
+    echo "-------------- file $wav_records/$wav_file BIGGERRRRRRRRR then $requiredsizewavfile "
+    echo "++++++++++++++ size of file  $wav_records/$wav_file   is $(stat -c%s $wav_records/$wav_file) "
+    lame -b $bitrate_to_mp3 $wav_records/$wav_file $wav_records/$mp3_file
+    mv -f $wav_records/$mp3_file $mp3_records/$mp3_file
+    aws s3 cp $mp3_records/$mp3_file $awss3bucket/$datepath/$mp3_file
+    rm -f $mp3_records/$mp3_file
+  fi
   echo "!!!!!!!!!!!!! will remove file =========== $wav_records/$wav_file   ======================"
+  echo "+++++++++++++ size of file  $wav_records/$wav_file   is $(stat -c%s $wav_records/$wav_file) "
   rm -f $wav_records/$wav_file
-  mv -f $wav_records/$mp3_file $mp3_records/$mp3_file
-  aws s3 cp $mp3_records/$mp3_file $awss3bucket/$mp3_file
-  rm -f $mp3_records/$mp3_file
 done
 echo "==============SHOW local files==============="
 echo "==============  Wav records  ==============="
@@ -62,6 +72,6 @@ ls -la $wav_records
 echo "==============  Mp3 records  ==============="
 ls -la $mp3_records
 echo "================ AWS S3 list ================"
-aws s3 ls s3://ugb-test-ast/
+aws s3 ls $awss3bucket/$datepath/
 #####################################################
 echo "///////////////////////--END SCRIPT--///////////////////////"
